+++ decidim-meetings/lib/decidim/meetings.rb
@@ -3,6 +3,8 @@
 require "decidim/meetings/admin"
 require "decidim/meetings/engine"
 require "decidim/meetings/admin_engine"
+require "decidim/meetings/directory"
+require "decidim/meetings/directory_engine"
 require "decidim/meetings/component"
 
 module Decidim